Use base class for Chademo battery
